gh-1346 - Specific validators implemented, new package for validators
