+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/function/FilterHandler.java
@@ -23,6 +23,8 @@ import uk.gov.gchq.gaffer.operation.util.StreamFilterIterable;
 import uk.gov.gchq.gaffer.store.Context;
 import uk.gov.gchq.gaffer.store.Store;
 import uk.gov.gchq.gaffer.store.operation.handler.OutputOperationHandler;
+import uk.gov.gchq.gaffer.store.operation.validator.function.FilterValidator;
+import uk.gov.gchq.gaffer.store.operation.validator.function.FunctionValidator;
 import uk.gov.gchq.gaffer.store.schema.Schema;
 import uk.gov.gchq.koryphe.ValidationResult;
 
@@ -31,7 +33,7 @@ import java.util.Map;
 
 public class FilterHandler implements OutputOperationHandler<Filter, Iterable<? extends Element>> {
 
-    private final FunctionValidator<Filter> validator = new FunctionValidator<>();
+    private final FunctionValidator<Filter> validator = new FilterValidator();
 
     @Override
     public Iterable<? extends Element> doOperation(final Filter operation, final Context context, final Store store) throws OperationException {
