+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/function/TransformHandler.java
@@ -22,10 +22,12 @@ import uk.gov.gchq.gaffer.store.Context;
 import uk.gov.gchq.gaffer.store.Store;
 import uk.gov.gchq.gaffer.store.operation.handler.OutputOperationHandler;
 import uk.gov.gchq.gaffer.store.operation.util.StreamTransformIterable;
+import uk.gov.gchq.gaffer.store.operation.validator.function.FunctionValidator;
+import uk.gov.gchq.gaffer.store.operation.validator.function.TransformValidator;
 import uk.gov.gchq.koryphe.ValidationResult;
 
 public class TransformHandler implements OutputOperationHandler<Transform, Iterable<? extends Element>> {
-    private final FunctionValidator<Transform> validator = new FunctionValidator<>();
+    private final FunctionValidator<Transform> validator = new TransformValidator();
 
     @Override
     public Iterable<? extends Element> doOperation(final Transform operation, final Context context, final Store store) throws OperationException {
